restructure LPC hooks
