+++ app/src/organisms/ProtocolSetup/LabwarePositionCheck/hooks/useCommands.ts
@@ -1,0 +1,11 @@
+import { LabwarePositionCheckCommand } from '../types'
+import { useSteps } from './useSteps'
+
+export function useCommands(): LabwarePositionCheckCommand[] {
+  return useSteps().reduce<LabwarePositionCheckCommand[]>(
+    (steps, currentStep) => {
+      return [...steps, ...currentStep.commands]
+    },
+    []
+  )
+}
